Replace JWT with API key (#211)
